requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
peewee==3.14.8
psycopg2==2.9.2
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
